server/requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7

# Redis
redis==5.0.1
Flask-Caching==2.1.0

# JWT Authentication
Flask-JWT-Extended==4.6.0

# CORS
Flask-Cors==4.0.0

# Email
Flask-Mail==0.9.1

# WebSocket
Flask-SocketIO==5.3.5
python-socketio==5.10.0
eventlet==0.33.3

# Password Hashing
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# HTTP Requests
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Image Processing (for profile pictures)
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-flask==1.3.0